requirements.txt
coverage==4.2
decorator==4.0.10
factory-boy==2.7.0
fake-factory==0.6.0
ipython==5.1.0
ipython-genutils==0.1.0
pexpect==4.2.0
pickleshare==0.7.3
prompt-toolkit==1.0.5
ptyprocess==0.5.1
py==1.4.31
Pygments==2.1.3
pytest==3.0.0
pytest-cov==2.3.1
python-dateutil==2.5.3
simplegeneric==0.8.1
six==1.10.0
SQLAlchemy==1.0.14
traitlets==4.2.2
wcwidth==0.1.7
